Refactor: Move Service Implementations to Infrastructure Layer
- Moved concrete application services (e.g., AuthService, BatchJobService) from Valora.Application/Services to Valora.Infrastructure/Services/AppServices to enforce Clean Architecture constraints.
- Moved interfaces located in Application/Services to Application/Common/Interfaces.
- Updated Dependency Injection registrations in both layers (removed scoped service registrations from Application, added them to Infrastructure).
- Fixed namespaces and deduplicated using statements across the Api, Application, Infrastructure, UnitTests, and IntegrationTests projects.
