+++ backend/Valora.Application/Common/Interfaces/IBatchJobProcessor.cs
@@ -1,7 +1,7 @@
 using Valora.Domain.Entities;
 using Valora.Domain.Enums;
 
-namespace Valora.Application.Services.BatchJobs;
+namespace Valora.Application.Common.Interfaces;
 
 public interface IBatchJobProcessor
 {
