+++ backend/Valora.Infrastructure/Services/AppServices/BatchJobs/CityIngestionJobProcessor.cs
@@ -5,7 +5,7 @@ using Valora.Domain.Entities;
 using Valora.Domain.Enums;
 using Valora.Domain.Extensions;
 
-namespace Valora.Application.Services.BatchJobs;
+namespace Valora.Infrastructure.Services.AppServices.BatchJobs;
 
 public class CityIngestionJobProcessor : IBatchJobProcessor
 {
