+++ backend/Valora.Infrastructure/Services/AppServices/NotificationService.cs
@@ -5,7 +5,7 @@ using Valora.Domain.Common;
 using Valora.Domain.Entities;
 using Valora.Domain.Enums;
 
-namespace Valora.Application.Services;
+namespace Valora.Infrastructure.Services.AppServices;
 
 public class NotificationService : INotificationService
 {
