+++ backend/Valora.IntegrationTests/BatchJobExecutorIntegrationTests.cs
@@ -9,7 +9,7 @@ using Moq;
 using Shouldly;
 using Valora.Application.Common.Events;
 using Valora.Application.Common.Interfaces;
-using Valora.Application.Services.BatchJobs;
+using Valora.Infrastructure.Services.AppServices.BatchJobs;
 using Valora.Domain.Entities;
 using Valora.Infrastructure.Persistence;
 using Xunit;
