+++ backend/Valora.IntegrationTests/ContextAnalysisSessionProfileTests.cs
@@ -4,7 +4,7 @@ using Microsoft.Extensions.DependencyInjection.Extensions;
 using Moq;
 using Valora.Application.Common.Interfaces;
 using Valora.Application.DTOs;
-using Valora.Application.Services;
+using Valora.Infrastructure.Services.AppServices;
 using Valora.Domain.Entities;
 using Valora.Infrastructure.Persistence;
 using Xunit;
