+++ backend/Valora.UnitTests/Services/AdminServiceTests.cs
@@ -3,7 +3,7 @@ using Moq;
 using Valora.Application.Common.Interfaces;
 using Valora.Application.Common.Models;
 using Valora.Application.DTOs;
-using Valora.Application.Services;
+using Valora.Infrastructure.Services.AppServices;
 using Valora.Domain.Entities;
 using Xunit;
