+++ backend/Valora.UnitTests/Services/BatchJobServiceGetJobsTests.cs
@@ -2,7 +2,7 @@ using Microsoft.Extensions.Logging;
 using Moq;
 using Valora.Application.Common.Interfaces;
 using Valora.Application.Common.Models;
-using Valora.Application.Services;
+using Valora.Infrastructure.Services.AppServices;
 using Valora.Domain.Entities;
 
 namespace Valora.UnitTests.Services;
