+++ backend/Valora.UnitTests/Services/BatchJobServiceTests.cs
@@ -2,8 +2,8 @@ using Microsoft.Extensions.Logging;
 using Moq;
 using Valora.Application.Common.Interfaces;
 using Valora.Application.DTOs;
-using Valora.Application.Services;
-using Valora.Application.Services.BatchJobs;
+using Valora.Infrastructure.Services.AppServices;
+using Valora.Infrastructure.Services.AppServices.BatchJobs;
 using Valora.Domain.Entities;
 
 namespace Valora.UnitTests.Services;
