+++ backend/Valora.UnitTests/Services/ContextReportScoringTests.cs
@@ -4,7 +4,7 @@ using Moq;
 using Valora.Application.Common.Interfaces;
 using Valora.Application.DTOs;
 using Valora.Application.Enrichment;
-using Valora.Application.Services;
+using Valora.Infrastructure.Services.AppServices;
 
 namespace Valora.UnitTests.Services;
