+++ backend/Valora.UnitTests/Services/Utilities/AmenityClustererTests.cs
@@ -1,5 +1,5 @@
 using Valora.Application.DTOs.Map;
-using Valora.Application.Services.Utilities;
+using Valora.Infrastructure.Services.AppServices.Utilities;
 using Xunit;
 
 namespace Valora.UnitTests.Services.Utilities;
